Replace pair dialog with dropdown
